# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/budgets:
    get:
      tags:
        - Budgets
      summary: List Budgets
      description: Get a paginated list of Budgets with optional filtering and sorting
      operationId: listBudgets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Budgets
      summary: Create Budget
      description: Create a new Budget
      operationId: createBudget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBudgetRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/budget_lines:
    get:
      tags:
        - BudgetLines
      summary: List BudgetLines
      description: Get a paginated list of BudgetLines with optional filtering and sorting
      operationId: listBudgetLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - BudgetLines
      summary: Create BudgetLine
      description: Create a new BudgetLine
      operationId: createBudgetLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBudgetLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Budget:
      type: object
      required:
        - id
        - company_id
        - code
        - name
        - fiscal_year
        - date_from
        - date_to
        - status
        - enforcement
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 100
        description:
          nullable: true
          type: string
          maxLength: 500
        fiscal_year:
          type: integer
          format: int32
          minimum: 2020
          maximum: 2100
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        status:
          type: string
        enforcement:
          type: string
        metadata:
          type: object
...
    CreateBudgetRequest:
      type: object
      required:
        - company_id
        - code
        - name
        - fiscal_year
        - date_from
        - date_to
        - status
        - enforcement
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 30
        name:
          type: string
          maxLength: 100
        description:
          nullable: true
          type: string
          maxLength: 500
        fiscal_year:
          type: integer
          format: int32
          minimum: 2020
          maximum: 2100
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        status:
          type: string
        enforcement:
          type: string
        metadata:
          type: object
...
    BudgetList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Budget'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    BudgetLine:
      type: object
      required:
        - id
        - budget_id
        - company_id
        - account_id
        - fiscal_period_id
        - fiscal_year
        - planned_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        budget_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        cost_center_id:
          nullable: true
          type: string
          format: uuid
        fiscal_period_id:
          type: string
          format: uuid
        fiscal_year:
          type: integer
          format: int32
          minimum: 2020
          maximum: 2100
        fiscal_month:
          nullable: true
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        planned_amount:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 500
        metadata:
          type: object
...
    CreateBudgetLineRequest:
      type: object
      required:
        - budget_id
        - company_id
        - account_id
        - fiscal_period_id
        - fiscal_year
        - planned_amount
        - metadata
      properties:
        budget_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        cost_center_id:
          nullable: true
          type: string
          format: uuid
        fiscal_period_id:
          type: string
          format: uuid
        fiscal_year:
          type: integer
          format: int32
          minimum: 2020
          maximum: 2100
        fiscal_month:
          nullable: true
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        planned_amount:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 500
        metadata:
          type: object
...
    BudgetLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BudgetLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []